Move custom spec helpers to their own directory
