+++ spec/support/helpers/stub_helper.rb
